Add number validation on lottery ticket
We need to check that the numbers assigned to a
ticket are exaclty 6 and they are between 1 and 49 inclusively.
We use custom model validations for the purpose.

A db constraint on that one would be best as we want data integrity
as close to the db as it gets. Good candidate for improvement.

I intentionally don't put such validation on the lottery draw model as
it generates numbers by itself (no user input) and don't feel like the
validation is needed. Probably another good candidate for improvement
is to have a module generate the six numbers instead of being method on the model
and extensively test that.